GH #2: remove unnecessary validation of _get_or_search:
In the method _get_or_search there's validation for a proper class
name. However, as described by rafl in the Github issue, it is both
incorrect and unnecessary.

There are both valid classes (for modules, distributions, and
releases) that won't pass the tests, and invalid classes which will
pass the test.

While it stands to reason to validate input before it is sent to
MetaCPAN, it is simply impractical to do this since the nature of
class names in Perl is... unfun to handle.

Therefor, we are only checking whether it's defined and not a
reference. If so, we pass it along to MetaCPAN. Enjoy.