Remove the duplicate "ua" attribute:
Instead of having two attributes, the BUILDARGS introduced will
create a MetaCPAN::Client::Request *always*, and will optionally
use a "ua" parameter sent to ->new().

Also, the "ua" method from MetaCPAN::Client::Request will be curried
with 'handles' on the attribute, so it can be used as if it's an
actual attribute of the MetaCPAN::Client object.