Role::Entity.pm: Add an optional, lazy loaded 'client' attribute.
This allows clients to create objects that remember who created them,
and thus allow created objects to call subsequent requests sharing the
Clients configuration options. ( Especially in regard to UA )

If no such attribute is passed, the it will be created on demand when
needed.

This will make existing code continue to function even if it doesn't use
the MetaCPAN Client base directly.