Removed Search::Elasticsearch
Created an internall scroller for the client.

We only use this one feature (scrolled searches) and we have it
as a dependency to function as an API to an API we could just
use directly.

The motivation is the breaking changes in latest 5.x versions of
Search::Elasticsearch and the fact it's incompatiable with the
current client code.