Fix GH #72: Buildling without '.' in @INC (5.26.0):
As kentfredric explains in GH #72, the missing dot in @INC (which
goes in 5.26.0) can break any test that uses t::lib::Functions,
assuming that it is available.

While this will work in a new Test::Harness, it wouldn't work for
people running it manually not under Test::Harness. (And we would
have to require that new version as a minimum dependency.)

Instead, this assures it is available when loading
t::lib::Functions.