Don't warn on scroller deletion (GH#81)
Failure to remove scroller will normally happen when the
scroller object goes out of scope after the server-side
scroller has timed-out and was removed.

Sending the DELETE is an act of politeness to help the
server cleaning up resources.

This warning doesn't add value to the user and apparently is
confusing.