Rewrite the scroller to stop skipping a page
On a large scroll over all of the "package" documents, the scroller was
implicated in consistently dropping a page of results.  The bookkeeping
by the scroller was a bit dodgy, so rather than figure out where the
off-by-one bug was, I just rewrote it.  Normally more rigor in
understanding the root cause is nice, but it's been a long day already
full of other hilarious bugs.