Test reverse dependencies:
This is a recreation of PR #90.

I have fixed up the styling and cleaned it up a bit. I couldn't find the
original repo that had this commit so I just copied it.

The original code was written by Steve Bertrand (@stevieb9).

I also removed the counting of the tests themselves because it seems to
be a fragile practice. We cannot control all of our reverse dependency
and any additional one uploaded to CPAN would break any existing
version until the counter is adjusted.

I would also note this test takes a few seconds to run. Any automatic CI
tsting would have to take that into account and monitored if it exceeds
the timeout for the entire testing suite.